WIP: Move models from rocketchat-lib to a specific package (server)
