+++ .meteor/packages
@@ -202,3 +202,4 @@ juliancwirko:postcss
 littledata:synced-cron
 rocketchat:utils
 rocketchat:settings
+rocketchat:models
