+++ packages/rocketchat-lib/package.js
@@ -25,6 +25,7 @@ Package.onUse(function(api) {
 	api.use('service-configuration');
 	api.use('check');
 	api.use('rocketchat:utils');
+	api.use('rocketchat:models');
 	api.use('rocketchat:accounts');
 	api.use('modules');
 	api.use('rocketchat:i18n');
