+++ packages/rocketchat-models/server/index.js
@@ -1,0 +1,27 @@
+import { Base } from './models/_Base';
+import { BaseDb } from './models/_BaseDb';
+import Avatars from './models/Avatars';
+import ExportOperations from './models/ExportOperations';
+import Messages from './models/Messages';
+import Reports from './models/Reports';
+import Rooms from './models/Rooms';
+import Settings from './models/Settings';
+import Subscriptions from './models/Subscriptions';
+import Uploads from './models/Uploads';
+import UserDataFiles from './models/UserDataFiles';
+import Users from './models/Users';
+
+export {
+	Base,
+	BaseDb,
+	Avatars,
+	ExportOperations,
+	Messages,
+	Reports,
+	Rooms,
+	Settings,
+	Subscriptions,
+	Uploads,
+	UserDataFiles,
+	Users,
+};
